manager/agent.py
from google.adk.agents import Agent
from .sub_agents.rag_agent.agent import rag_agent
from .sub_agents.sop_agent.agent import sop_agent
from .sub_agents.adk_bq_agent.agent import adk_bq_agent

root_agent = Agent(
    name="manager",
    model="gemini-2.5-flash-lite",
    description="A manager agent that delegates tasks to sub-agents for RAG (Retrieval-Augmented Generation), SOP (Standard Operating Procedure) queries, and SQL queries.",
    sub_agents=[
        rag_agent,
        sop_agent,
        adk_bq_agent,
    ],
    instruction="""
    You are the manager agent. Your primary role is to understand the user's request and delegate it to the appropriate sub-agent. You have three sub-agents available:

    1.  **RAG_agent**: Use this agent for any questions related to creating, managing, or querying document corpora for Retrieval-Augmented Generation (RAG). This includes adding data, creating or deleting corpora, and performing RAG queries.

    2.  **SOP_agent**: Use this agent for questions about 'Those Who Care (TWC)' Standard Operating Procedures (SOPs). This agent can find SOPs and answer questions about specific procedures.

    3.  **adk_bq_agent**: Use this agent for any questions related to querying databases using SQL. This includes running queries on BigQuery tables.

    Carefully analyze the user's query to determine which agent is best suited to handle the task.
    - If the query involves "RAG", "corpus", "document", "knowledge base", or "information retrieval", use the `rag_agent`.
    - If the query involves "SOP", "procedure", "how to", "process", or asks for instructions on a task at TWC, use the `sop_agent`.
    - If the query involves "SQL", "database", "query", "table", or "BigQuery", use the `adk_bq_agent`.

    If the query is ambiguous, ask for clarification before delegating. If the query cannot be handled by any of the sub-agents, inform the user of your limitations.
    """,
)
