requirements.txt
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
google-api-python-client
gspread
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv

# google-cloud-aiplatform[agent_engines, adk]
# google-adk
cloudpickle
pydantic
pyyaml
ipykernel
# python-dotenv
google-cloud-bigquery